added JWT auth functionality

+++ app/Http/Controllers/API/ContactsController.php
@@ -6,50 +6,37 @@ use App\Http\Requests;
 use App\Models\Contact;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use Auth;
 use JWTAuth;
 use Validate;
 
 class ContactsController extends Controller
 {
     /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        $this->middleware('auth');
-    }
-
-    /**
      * Show the application dashboard.
      *
      * @return Response
      */
 
+    // Has $id = NULL for now - can be removed (and in routes.php) if I end up
+    // Not using it.
     public function get(Request $request, $id = NULL)
     {
-        $user = Auth::user();
+        $user = JWTAuth::parseToken()->authenticate();
 
-        // if(!$this->_authenticate()) {
-        //     echo "Not authenticated ...";
-        // }
+        if(!$user) {
+            return response()->json('Not logged in', 401);
+        }
 
-        if(isset($id))
-        {
-            $contact = $user->contacts->where('id', $id);
-        }
-        else
-        {
-            $contacts = $user->contacts;
-            return response()->json($contacts);
-        }
+        return response()->json($user->contacts);
     }
 
     public function insert(Request $request)
     {
-        $user = Auth::user();
+        $user = JWTAuth::parseToken()->authenticate();
+
+        if(!$user) {
+            return response()->json('Not logged in');
+        }
 
         $this->validate($request, [
                 'name' => 'required|max:255',
@@ -74,19 +61,32 @@ class ContactsController extends Controller
 
     public function delete(Request $request, $id)
     {
+        $user = JWTAuth::parseToken()->authenticate();
+
+        if(!$user) {
+            return response()->json('Not logged in');
+        }
+
+        //ToDo: check if contact is really one of authenticated user's contacts
+
         $result = Contact::destroy($id);
         return response()->json($result);
     }
 
     public function update(Request $request)
     {
-        $user = Auth::user();
+        $user = JWTAuth::parseToken()->authenticate();
+
+        if(!$user) {
+            return response()->json('Not logged in');
+        }
 
         $this->validate($request, [
                 'name' => 'required|max:255',
                 'number' => 'required|max:20|min:6'
             ]);
 
+        //ToDo: check if contact is really one of authenticated user's contacts
         $contact = Contact::find($request->id);
         $contact->name = $request->name;
         $contact->number = $request->number;
@@ -94,10 +94,5 @@ class ContactsController extends Controller
         $result = $contact->save();
 
         return response()->json($result);
-    }
-
-    private function _authenticate()
-    {
-        return JWTAuth::parseToken()->authenticate();
     }
 }
